+++ ndsolve.h
@@ -1,9 +1,9 @@
 #include <math.h>
 
-double a_x(double r, double x);
-double a_y(double r, double y);
+float a_x(float r, float x);
+float a_y(float r, float y);
 
-void solve_ode(unsigned short offset, unsigned short steps, double h, double x[steps], double y[steps], double vx[steps], double vy[steps], double ax[steps], double ay[steps]);
+void solve_ode(unsigned short offset, unsigned short steps, float h, float x[steps], float y[steps], float vx[steps], float vy[steps], float ax[steps], float ay[steps]);
 
-void adams_bashford(unsigned short offset, unsigned short steps, double h, double x[steps], double y[steps], double vx[steps], double vy[steps], double ax[steps], double ay[steps]);
+void adams_bashford(unsigned short offset, unsigned short steps, float h, float x[steps], float y[steps], float vx[steps], float vy[steps], float ax[steps], float ay[steps]);
